+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added the `exists` filter to check if a variable with some name exists in the current scope
 - Added the `value` filter to get the value of a variable specified by a name in the current scope or a default
+- Added the `apply_mapping` filter to map objects or variables in scope to a new object
+- Added a `fatality` filter to exit template processing with an error message
 
 ### Changed
 - Updated informatory texts